Move BumpVersion to models from config
