+++ changes/commands/stage.py
@@ -10,7 +10,7 @@ from jinja2 import Template
 import changes
 from changes.config import BumpVersion
 
-from changes.models import Release, changes_to_release_type
+from changes.models import Release, changes_to_release_type, BumpVersion
 from . import info, error, debug, STYLES
